functions/vlr-stats-scrapper/requirements.txt
google-cloud-logging>=3.13.0
google-cloud-storage>=3.9.0
httpx>=0.28.1
selectolax>=0.4.6